+++ src/repository/channel.ts
@@ -168,7 +168,7 @@ export class ChannelsRepository implements IRepository<YtChannel> {
 
   // lock any updates on video table
   private readonly ASYNC_LOCK_ID = 'channel'
-  private asyncLock: AsyncLock = new AsyncLock()
+  private asyncLock: AsyncLock = new AsyncLock({ maxPending: Number.MAX_SAFE_INTEGER })
 
   constructor(tablePrefix: ResourcePrefix) {
     this.model = createChannelModel(tablePrefix)