+++ src/utils/configParser.ts
@@ -96,8 +96,8 @@ export class ConfigParserService {
               _.set(config, path, envValue)
           }
         }
-        const errors = this.validator.errorsByProperty('Config', path.join('.'), config)
-        if (errors) {
+        const errors = this.validator.errorsByProperty('Config', path.join('/'), config)
+        if (errors?.length) {
           throw new ValidationError(`Invalid env value`, errors)
         }
         return