+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### 1.4.2
 
-- Add support for transcoding the media files while downloading & merging the audio/video streams from Youtube. E.g, If the media output container is `mkv` then transcode(recode) the media to `mp4`, as `mkv` is not widely supported by the browsers.
+- Adds Youtube API quota usage optimization.
+- Fetches Youtube API daily quota usage stats from the Google's Cloud Monitoring API instead of calculating it in the application.
+- Bumps `ajv` package version to `^8.0.0` & generate docs for the configuration schema
+- set AsyncLock's queue size to `MAX_SAFE_INTEGER`
 
 ### 1.4.1